Correcion prediccion futuro

+++ BolsaPython/bolsa/C6CreadorModelosDeSubgrupo.py
@@ -1,5 +1,4 @@
 import sys
-import os
 import pandas as pd
 import numpy as np
 np.random.seed(12345)
@@ -18,6 +17,7 @@ from sklearn.utils import resample
 from sklearn.model_selection import GridSearchCV
 from sklearn.calibration import CalibratedClassifierCV
 from shutil import copyfile
+import os.path
 
 print("-------------------------------------------------------------------------------------------------")
 print("---- CAPA 6 - Crear almacenar y evaluar varios modelos (para cada subgrupo) -------")
@@ -303,52 +303,63 @@ elif (modoTiempo == "futuro" and pathCsvReducido.endswith('.csv') and os.path.is
         if file.endswith("ganador"):
             path_modelo_predictor_ganador = os.path.join(dir_modelo_predictor_ganador, file)
 
-    print("Cargar modelo PREDICTOR ganador (de la carpeta del pasado, claro): " + path_modelo_predictor_ganador)
-    modelo_predictor_ganador = pickle.load(open(path_modelo_predictor_ganador, 'rb'))
+    print("Cargar modelo PREDICTOR ganador (de la carpeta del pasado, SI EXISTE): " + path_modelo_predictor_ganador)
+    if os.path.isfile(path_modelo_predictor_ganador):
 
-    print("Predecir:")
-    targets_predichos = modelo_predictor_ganador.predict(inputFeatures_sinnulos)
-    print("Numero de targets_predichos: " + str(targets_predichos.shape[0]))
-    print("Guardando targets PREDICHOS en: " + pathCsvPredichos)
-    pathCsvPredichos_df = pd.DataFrame(data=targets_predichos, columns=['TARGET'])
-    pathCsvPredichos_df.to_csv(pathCsvPredichos, index=False, sep='|')
+        modelo_predictor_ganador = pickle.load(open(path_modelo_predictor_ganador, 'rb'))
 
-    print("Guardando indices de filas de salida respecto de la entrada...")
-    np.savetxt(pathCsvPredichos + "_indices", indiceFilasFuturasTransformadas2, delimiter="|", header="indice", fmt='%f')
+        print("Predecir:")
+        targets_predichos = modelo_predictor_ganador.predict(inputFeatures_sinnulos)
+        print("Numero de targets_predichos: " + str(targets_predichos.shape[0]))
+        print("Guardando targets PREDICHOS en: " + pathCsvPredichos)
+        pathCsvPredichos_df = pd.DataFrame(data=targets_predichos, columns=['TARGET'])
+        pathCsvPredichos_df.to_csv(pathCsvPredichos, index=False, sep='|')
 
+        print("Guardando indices de filas de salida respecto de la entrada...")
+        np.savetxt(pathCsvPredichos + "_indices", indiceFilasFuturasTransformadas2, delimiter="|", header="indice",
+                   fmt='%f')
 
-    ############### RECONSTRUCCION DEL CSV FINAL IMPORTANTE, viendo los ficheros de indices #################
-    print("Partiendo de COMPLETO.csv llevamos la cuenta de los indices pasando por REDUCIDO.csv y por TARGETS_PREDICHOS.csv para generar el CSV final...")
-    df_completo = pd.read_csv(pathCsvCompleto, sep='|')
-    df_reducido = pd.read_csv(pathCsvReducido, sep='|')
-    df_reducido_indices = pd.read_csv(pathCsvReducidoIndices, sep='|').to_numpy()
-    df_predichos = pd.read_csv(pathCsvPredichos, sep='|')
-    df_predichos_indices = pd.read_csv(pathCsvPredichosIndices, sep='|').to_numpy()
+        ############### RECONSTRUCCION DEL CSV FINAL IMPORTANTE, viendo los ficheros de indices #################
+        print(
+            "Partiendo de COMPLETO.csv llevamos la cuenta de los indices pasando por REDUCIDO.csv y por TARGETS_PREDICHOS.csv para generar el CSV final...")
+        df_completo = pd.read_csv(pathCsvCompleto, sep='|')
+        df_reducido = pd.read_csv(pathCsvReducido, sep='|')
+        df_reducido_indices = pd.read_csv(pathCsvReducidoIndices, sep='|').to_numpy()
+        df_predichos = pd.read_csv(pathCsvPredichos, sep='|')
+        df_predichos_indices = pd.read_csv(pathCsvPredichosIndices, sep='|').to_numpy()
 
-    print("df_completo: " + str(df_completo.shape[0]) + " x " + str(df_completo.shape[1]))
-    print("df_reducido: " + str(df_reducido.shape[0]) + " x " + str(df_reducido.shape[1]))
-    print("df_reducido_indices: " + str(df_reducido_indices.shape[0]) + " x " + str(df_reducido_indices.shape[1]))
-    print("df_predichos: " + str(df_predichos.shape[0]) + " x " + str(df_predichos.shape[1]))
-    print("df_predichos_indices: " + str(df_predichos_indices.shape[0]) + " x " + str(df_predichos_indices.shape[1]))
+        print("df_completo: " + str(df_completo.shape[0]) + " x " + str(df_completo.shape[1]))
+        print("df_reducido: " + str(df_reducido.shape[0]) + " x " + str(df_reducido.shape[1]))
+        print("df_reducido_indices: " + str(df_reducido_indices.shape[0]) + " x " + str(df_reducido_indices.shape[1]))
+        print("df_predichos: " + str(df_predichos.shape[0]) + " x " + str(df_predichos.shape[1]))
+        print(
+            "df_predichos_indices: " + str(df_predichos_indices.shape[0]) + " x " + str(df_predichos_indices.shape[1]))
 
-    print("COMPLETO_1 --> Del COMPLETO, cogemos solo las filas que se han usado para calcular el REDUCIDO...")
-    df_completo_1 = df_completo.loc[df_reducido_indices[:,0].tolist()].reset_index(drop=False, inplace=False).drop('index', axis=1).drop('TARGET', axis=1)
-    print("df_completo_1: " + str(df_completo_1.shape[0]) + " x " + str(df_completo_1.shape[1]))
+        print("COMPLETO_1 --> Del COMPLETO, cogemos solo las filas que se han usado para calcular el REDUCIDO...")
+        df_completo_1 = df_completo.loc[df_reducido_indices[:, 0].tolist()].reset_index(drop=False, inplace=False).drop(
+            'index', axis=1).drop('TARGET', axis=1)
+        print("df_completo_1: " + str(df_completo_1.shape[0]) + " x " + str(df_completo_1.shape[1]))
 
-    print("COMPLETO_2 --> Del COMPLETO_1, cogemos solo las filas que se han usado para calcular el PREDICHOS...")
-    df_completo_2 = df_completo_1.loc[df_predichos_indices[:, 0].tolist()]
-    print("df_completo_2: " + str(df_completo_2.shape[0]) + " x " + str(df_completo_2.shape[1]))
+        print("COMPLETO_2 --> Del COMPLETO_1, cogemos solo las filas que se han usado para calcular el PREDICHOS...")
+        df_completo_2 = df_completo_1.loc[df_predichos_indices[:, 0].tolist()]
+        print("df_completo_2: " + str(df_completo_2.shape[0]) + " x " + str(df_completo_2.shape[1]))
 
-    print("Juntar COMPLETO con TARGETS PREDICHOS... ")
-    df_juntos_1 = pd.concat([df_completo_2.reset_index(drop=True), df_predichos], axis=1)
-    lado_derecho=df_juntos_1[['empresa', 'mercado', 'antiguedad', 'TARGET']]
-    df_juntos_2 = pd.merge(df_completo, lado_derecho, how='left', on=['empresa', 'mercado', 'antiguedad'])
-    df_juntos_2.rename(columns={'TARGET_x': 'TARGET_REAL', 'TARGET_y': 'TARGET_PREDICHO'}, inplace=True)
+        print("Juntar COMPLETO con TARGETS PREDICHOS... ")
+        df_juntos_1 = pd.concat([df_completo_2.reset_index(drop=True), df_predichos], axis=1)
+        lado_derecho = df_juntos_1[['empresa', 'mercado', 'antiguedad', 'TARGET']]
+        df_juntos_2 = pd.merge(df_completo, lado_derecho, how='left', on=['empresa', 'mercado', 'antiguedad'])
+        df_juntos_2.rename(columns={'TARGET_x': 'TARGET_REAL', 'TARGET_y': 'TARGET_PREDICHO'}, inplace=True)
 
-    df_juntos_2['TARGET_PREDICHO'] = (df_juntos_2['TARGET_PREDICHO'] * 1).astype('Int64') # Convertir de boolean a int64, manteniendo los nulos
+        df_juntos_2['TARGET_PREDICHO'] = (df_juntos_2['TARGET_PREDICHO'] * 1).astype(
+            'Int64')  # Convertir de boolean a int64, manteniendo los nulos
 
-    print("Guardando: " + pathCsvFinalFuturo)
-    df_juntos_2.to_csv(pathCsvFinalFuturo, index=False, sep='|')
+        print("Guardando: " + pathCsvFinalFuturo)
+        df_juntos_2.to_csv(pathCsvFinalFuturo, index=False, sep='|')
+                
+
+    else:
+        print("No existe el modelo predictor del pasado que necesitamos (" + path_modelo_predictor_ganador + "). Por tanto, no predecimos.")
+
 
 else:
     print("Los parametros de entrada son incorrectos o el CSV no existe o esta vacio!!")
